feat[UWSM]: migrate to uwsm for sandboxing and better systemd application management

+++ .config/rofi/scripts/power-menu.sh
@@ -6,19 +6,23 @@ if [ "${ROFI_RETV:-}" = "1" ]; then
     case "${ROFI_INFO:-}" in
         poweroff)
             systemctl poweroff
-            ;;
+        ;;
         reboot)
             systemctl reboot
-            ;;
+        ;;
         logout)
-            hyprctl dispatch 'hl.dsp.exit()' || pkill -x Hyprland
-            ;;
+            if command -v uwsm >/dev/null 2>&1 && uwsm check; then
+                uwsm stop
+            else
+                hyprctl dispatch 'hl.dsp.exit()' || pkill -x Hyprland
+            fi
+        ;;
         sleep)
             systemctl suspend
-            ;;
+        ;;
         lock)
             hyprlock -c ~/.config/hypr/hyprlock.conf
-            ;;
+        ;;
     esac
     exit 0
 fi
@@ -29,5 +33,4 @@ printf 'sleep\0info\x1fsleep\n'
 printf 'reboot\0info\x1freboot\n'
 printf 'poweroff\0info\x1fpoweroff\n'
 printf 'logout\0info\x1flogout\n'
-
 printf '\0message\x1fpower\n'
